requirements.txt
django==1.7.1
MySQL-python==1.2.5
django-decouple==2.1
